Staff filter shouldn't search nearby y values
Previously, the staff filter refined the y-value estimate, and only
kept the single-pixel refined line. However, this undermines the Hough
transform since we could actually have a rhores-wide line.
Therefore, it made the center line less prominent, and the line above
or below it could be detected if there are prominent ledger lines.